+++ src/operators/kernel/fpga/elementwise_add_relu_kernel.cpp
@@ -25,10 +25,10 @@ bool ElementwiseAddReluKernel<FPGA, float>::Init(
   Tensor *input_x = const_cast<Tensor *>(param->InputX());
   Tensor *input_y = const_cast<Tensor *>(param->InputY());
   Tensor *out = param->Out();
-  auto input_x_ptr = input_x->data<half>();
-  auto input_y_ptr = input_y->data<half>();
+  auto input_x_ptr = input_x->data<float>();
+  auto input_y_ptr = input_y->data<float>();
   fpga::format_ofm(out);
-  auto out_ptr = out->mutable_data<half>();
+  auto out_ptr = out->mutable_data<float>();
 
   fpga::EWAddArgs ewaddArgs;
   ewaddArgs.relu_enabled = relu_enabled;
