+++ src/operators/kernel/fpga/pool_kernel.cpp
@@ -22,10 +22,10 @@ namespace operators {
 template <>
 bool PoolKernel<FPGA, float>::Init(PoolParam<FPGA> *param) {
   Tensor *input = const_cast<Tensor *>(param->Input());
-  auto input_ptr = input->data<half>();
+  auto input_ptr = input->data<float>();
   Tensor *output = param->Output();
   fpga::format_ofm(output);
-  auto output_ptr = output->mutable_data<half>();
+  auto output_ptr = output->mutable_data<float>();
   vector<int> ksize = param->Ksize();
   vector<int> strides = param->Strides();
   vector<int> paddings = param->Paddings();
